fixup! refactor(proxy): Move binary code into `api` crate
