+++ proxy/api/src/main.rs
@@ -1,4 +1,13 @@
 #[tokio::main]
 pub async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    api::main().await
+    api::env::set_if_unset("RUST_BACKTRACE", "full");
+    api::env::set_if_unset("RUST_LOG", "info");
+    pretty_env_logger::init();
+
+    let mut args = pico_args::Arguments::from_env();
+    let args = api::Args {
+        test: args.contains("--test"),
+    };
+
+    api::run(args).await
 }
